Enter

+++ TypeConversionPrimitive/ExplicitConversion.java
@@ -2,7 +2,7 @@ package TypeConversionPrimitive;
 
 public class ExplicitConversion {
      public static void main(String[] args) {
-    	 double d = 65.4;
+    	 float d = 65.4f;
     	 int i = (int)d;
     	 char c = (char)i;
     	 System.out.println(i);
